requirements.txt
asgiref==3.4.1
Django==4.0.1
djangorestframework==3.13.1
psycopg2==2.9.3
Pygments==2.11.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
pillow
django-createsuperuser